requirements.txt
anyio==4.6.2.post1
base58==2.1.1
certifi==2024.8.30
charset-normalizer==3.4.0
construct==2.10.68
construct-typing==0.5.6
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonalias==0.1.1
numpy==2.1.3
opencv-python-headless==4.10.0.84
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
solana==0.35.1
solders==0.21.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==11.0.3